+++ src/search_store$_/index.d.ts
@@ -1,7 +1,7 @@
-import { ReadableAtom$ } from '@ctx-core/nanostores'
-export declare function search_result_store_<I extends unknown = unknown, O extends unknown = unknown>({ query$, data_, clear, timeout }:search_result_opts_I<I, O>):search_result$_T<I, O>;
+import { ReadableAtom, ReadableAtom$ } from '@ctx-core/nanostores'
+export declare function search_store$_<I extends unknown = unknown, O extends unknown = unknown>({ query$, data_, clear, timeout }:search_result_opts_I<I, O>):search_result$_T<I, O>;
 export interface search_result_opts_I<I extends unknown = unknown, O extends unknown = unknown> {
-	query$:ReadableAtom$<I>;
+	query$:ReadableAtom<I>;
 	data_:(params:{
 		query:I;
 	})=>Promise<O[]>;
@@ -11,9 +11,12 @@ export interface search_result_opts_I<I extends unknown = unknown, O extends unk
 export interface search_result_T<I extends unknown = unknown, O extends unknown = unknown> {
 	done:boolean;
 	loading:boolean;
-	query$:ReadableAtom$<I>;
+	query:I;
 	data?:O[];
 }
 export interface search_result$_T<I extends unknown = unknown, O extends unknown = unknown> extends ReadableAtom$<search_result_T<I, O>> {
 }
-export { search_result_store_ as _store__search_result }
+export {
+	search_store$_ as search_result_store_,
+	search_store$_ as _store__search_result,
+}