+++ src/search_store$_/index.js
@@ -1,27 +1,26 @@
 import { run } from '@ctx-core/function'
 import { setter_computed$ } from '@ctx-core/nanostores'
-export function search_result_store_({ query$, data_, clear, timeout }) {
+export function search_store$_({ query$, data_, clear, timeout }) {
 	let current_search_store
 	const search_store$ = setter_computed$(query$, (query, _set)=>{
-		// const set = in_set as (value:search_result_T)=>void
 		if (!query) {
 			run(clear || (()=>{
-				set(_set, {
+				set({
 					done: true,
 					loading: false,
-					query$,
+					query: query$.get(),
 					data: []
 				})
 			}))
 			return
 		}
-		if ((current_search_store === null || current_search_store === void 0 ? void 0 : current_search_store.query$.$) === query) {
+		if (current_search_store?.query === query) {
 			return
 		}
-		set(_set, {
+		set({
 			done: false,
 			loading: true,
-			query$
+			query: query$.get()
 		})
 		if (typeof typeof timeout !== 'number') timeout = 10000
 		Promise.race([
@@ -30,24 +29,26 @@ export function search_result_store_({ query$, data_, clear, timeout }) {
 					, timeout)
 			}),
 			run(async ()=>{
-				const data = await data_({
-					query
-				})
-				if (query === query$.$) {
-					set(_set, {
+				const data = await data_({ query })
+				const $query$ = query$.get()
+				if (query === $query$) {
+					set({
 						done: true,
 						loading: false,
-						query$,
+						query: $query$,
 						data
 					})
 				}
 			})
 		]).then()
+		function set(search_store) {
+			current_search_store = search_store
+			_set(search_store)
+		}
 	})
 	return search_store$
-	function set(_set, search_store) {
-		current_search_store = search_store
-		_set(search_store)
-	}
 }
-export { search_result_store_ as _store__search_result }
+export {
+	search_store$_ as search_result_store_,
+	search_store$_ as _store__search_result
+}